pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "business-rules"
version = "0.1.0"
description = "Rule-based fraud detection engine with LLM explanations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pandas>=2.1.0",
    "faker>=22.0.0",
    "anthropic>=0.42.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "websockets>=13.0",
]
notebook = [
    "jupyter>=1.1.0",
    "ipykernel>=6.29.0",
]
dev = [
    "black>=24.10.0",
    "ruff>=0.8.6",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
